# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pinjie Fullstack Blog Backend",
    "description": "独立个人博客的后端 API，当前提供认证、账户、后台管理、文件资产和系统设置基础能力。",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "用户认证"
        ],
        "summary": "注册用户账户",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserAuthSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "用户认证"
        ],
        "summary": "用户登录",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserAuthSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "用户认证"
        ],
        "summary": "刷新用户登录会话",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "pinjie_blog_web_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RefreshSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "用户认证"
        ],
        "summary": "退出当前用户会话",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "pinjie_blog_web_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_bool_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "管理员认证"
        ],
        "summary": "管理员登录",
        "operationId": "login_api_v1_admin_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminAuthSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/refresh": {
      "post": {
        "tags": [
          "管理员认证"
        ],
        "summary": "刷新管理员登录会话",
        "operationId": "refresh_api_v1_admin_auth_refresh_post",
        "parameters": [
          {
            "name": "pinjie_blog_admin_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RefreshSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/logout": {
      "post": {
        "tags": [
          "管理员认证"
        ],
        "summary": "退出当前管理员会话",
        "operationId": "logout_api_v1_admin_auth_logout_post",
        "parameters": [
          {
            "name": "pinjie_blog_admin_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_bool_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminAuthSessionOut": {
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/AdminRead",
            "description": "当前认证主体信息"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "登录会话唯一标识"
          },
          "access_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Access Expires At",
            "description": "访问凭据过期时间"
          },
          "idle_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Idle Expires At",
            "description": "会话空闲过期时间"
          },
          "absolute_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Absolute Expires At",
            "description": "会话绝对过期时间"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "principal",
          "session_id",
          "access_expires_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "AdminAuthSessionOut"
      },
...
      "AdminLoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Password",
            "description": "登录密码，最多 64 个字符"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "AdminLoginIn"
      },
...
      "AdminRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "资源唯一标识"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "展示名称"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar",
            "description": "管理员头像 URL 或站内资源路径"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "资源当前是否启用"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "description": "管理员是否拥有超级管理员身份"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            },
            "type": "array",
            "title": "Roles",
            "description": "管理员当前拥有的角色列表"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "当前主体拥有的权限代码列表"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "最近更新时间"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "is_active",
          "is_superuser",
          "roles",
          "permissions",
          "created_at",
          "updated_at"
        ],
        "title": "AdminRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail",
            "description": "请求参数校验错误详情列表"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefreshSessionOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "登录会话唯一标识"
          },
          "access_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Access Expires At",
            "description": "访问凭据过期时间"
          },
          "idle_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Idle Expires At",
            "description": "会话空闲过期时间"
          },
          "absolute_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Absolute Expires At",
            "description": "会话绝对过期时间"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "access_expires_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "RefreshSessionOut"
      },
...
      "ResponseModel_AdminAuthSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/AdminAuthSessionOut",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminAuthSessionOut]"
      },
...
      "ResponseModel_RefreshSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/RefreshSessionOut",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[RefreshSessionOut]"
      },
...
      "ResponseModel_UserAuthSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/UserAuthSessionOut",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[UserAuthSessionOut]"
      },
...
      "ResponseModel_bool_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "type": "boolean",
            "title": "Data",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[bool]"
      },
...
      "RoleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "资源唯一标识"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "资源名称"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "RoleSummary"
      },
...
      "UserAuthSessionOut": {
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/UserPrincipalOut",
            "description": "当前认证主体信息"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "登录会话唯一标识"
          },
          "access_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Access Expires At",
            "description": "访问凭据过期时间"
          },
          "idle_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Idle Expires At",
            "description": "会话空闲过期时间"
          },
          "absolute_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Absolute Expires At",
            "description": "会话绝对过期时间"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "principal",
          "session_id",
          "access_expires_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "UserAuthSessionOut"
      },
...
      "UserLoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Password",
            "description": "登录密码，最多 64 个字符"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLoginIn"
      },
      "UserPrincipalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "资源唯一标识"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "展示名称"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱地址"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar",
            "description": "用户头像站内资源路径"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "资源当前是否启用"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "最近更新时间"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "email",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "UserPrincipalOut"
      },
      "UserRegisterIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "title": "Password",
            "description": "登录密码，长度为 6 至 64 个字符"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "展示名称"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱地址"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserRegisterIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location",
            "description": "错误字段在请求中的位置"
          },
          "msg": {
            "type": "string",
            "title": "Message",
            "description": "参数校验错误消息"
          },
          "type": {
            "type": "string",
            "title": "Error Type",
            "description": "参数校验错误类型"
          },
          "input": {
            "title": "Input",
            "description": "引发校验错误的输入值，敏感内容可能被省略"
          },
          "ctx": {
            "type": "object",
            "title": "Context",
            "description": "参数校验错误的补充上下文"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}